fix:提交空信息bug

+++ src/components/bodyFatRate.vue
@@ -3,29 +3,40 @@
     <div class="back" @click="back">上一题</div>
     <h3>请输入您的体脂率</h3>
     <div class="form">
-      <input type="number" placeholder="请输入体脂率" v-model="bodyFatRate"/> <strong>%</strong>
-      <button @click="confirm">确认</button>
+      <input type="number" placeholder="请输入体脂率" v-model="bodyFatRate" @input="inputFn"/>
+      <strong>%</strong>
+      <button @click="confirm" :class="{ active: isActive }">确认</button>
     </div>
   </div>
 </template>
 
 <script>
 export default {
-  data(){
-    return{
-      bodyFatRate:null
+  data() {
+    return {
+      bodyFatRate: null,
+      isActive: false,
     }
   },
   methods: {
     back() {
       this.$router.go(-1)
     },
-    confirm(){
-      let user = JSON.parse(localStorage.getItem("userInfo"))
-      let obj = {...user,"bodyFatRate": this.bodyFatRate/1}
-      localStorage.setItem("userInfo",JSON.stringify(obj));
-      this.$router.push('/waistline')
-    }
+    inputFn() {
+      if (this.bodyFatRate) {
+        this.isActive = true
+      }else{
+        this.isActive = false
+      }
+    },
+    confirm() {
+      if (this.bodyFatRate){
+        let user = JSON.parse(localStorage.getItem('userInfo'))
+        let obj = { ...user, bodyFatRate: this.bodyFatRate / 1 }
+        localStorage.setItem('userInfo', JSON.stringify(obj))
+        this.$router.push('/waistline')
+      }
+    },
   },
 }
 </script>
@@ -49,20 +60,23 @@ export default {
       padding: 5px;
       text-align: center;
       background-color: #f5f5f5;
+      margin-right: 4px;
     }
   }
-  button{
+  button {
     width: 120px;
     height: 50px;
-    background-color: #fdf052;
     border-radius: 20px;
     border: none;
     text-align: center;
     font-weight: 600;
-    display:block;
+    display: block;
     margin-left: auto;
     margin-right: auto;
     margin-top: 24px;
+  }
+  .active {
+    background-color: #fdf052;
   }
 }
 </style>
